refactor(tests): Categories service - 5 files + helpers, 105 tests passing
